+++ tests/test_harvesting.py
@@ -69,7 +69,7 @@ class SimpleTest(TestCase):
     def test_api_root(self):
         response = self.client.get("/api/v1/works/")
         self.assertEqual(response.status_code, 200)
-        self.assertEqual(response.get("Content-Type"), "application/json")
+        self.assertEqual(response.get("Content-Type"), "application/geo+json")
         results = response.json()["results"]
         self.assertEqual(results["type"], "FeatureCollection")
         self.assertEqual(len(results["features"]), 2)